Withdraw

+++ controllers/ldap.py
@@ -16,7 +16,7 @@ def connect_ldap_server():
         connection = e
     return connection
 
-def add_group(user_id):
+def add_group(user_id, passwd):
     ldap_attr = {}
     ldap_attr['objectClass'] = ['top', 'posixGroup']
     new_gid = config.LDAP_NEW_GID
@@ -41,7 +41,7 @@ def add_group(user_id):
         if isSuccess == False:
             # 사용자가 회원가입을 동일하게 재진행할 수 있도록 LDAP의 group, user 삭제
             delete_group(user_id)
-            delete_user(user_id)
+            delete_user(user_id, passwd)
             return False
         kc_group.post_project_name(user_id)
         ks_project.post_project_id(user_id)
@@ -73,7 +73,7 @@ def add_user(user_id, user_name, mail, passwd, isUser):
             return "user"
 
         time.sleep(0.6)
-        isSuccess = add_group(user_id)
+        isSuccess = add_group(user_id, passwd)
         if isSuccess == 409:
             return 409
 
@@ -82,13 +82,13 @@ def add_user(user_id, user_name, mail, passwd, isUser):
         if isSuccess == False:
             # 사용자가 회원가입을 동일하게 재진행할 수 있도록 LDAP의 group, user 삭제
             delete_group(user_id)
-            delete_user(user_id)
+            delete_user(user_id, passwd)
             return False
         
         isSuccess = kc_user.post_user_attributes(user_id, isUser)
         if isSuccess == False:
             delete_group(user_id)
-            delete_user(user_id)
+            delete_user(user_id, passwd)
             return False           
 
     except LDAPException as e:
@@ -190,16 +190,22 @@ def delete_group(user_id):
            response = e
     return response
 
-def delete_user(user_id):
+def delete_user(user_id, password):
     ldap_conn = connect_ldap_server()
     user_dn = "cn="+user_id+",ou=users,cn=admin,dc=devstack,dc=co,dc=kr"
-
+    
+    # 사용자 존재 여부 확인
     check = check_user(user_id)
     if len(check) == 0:
         return "user"
 
+    # 비밀번호 검증
+    check = check_password_validation(user_id, password)
+    if check == False:
+        return "passwd"
+
+    ks_project.delete_project(user_id)
     delete_group(user_id)
-    #ks_project.delete_project(user_name)
     ks_user.delete_user(user_id)
 
     try:
@@ -207,3 +213,16 @@ def delete_user(user_id):
     except LDAPException as e:
         response = e
     return response
+
+
+def check_password_validation(user_id, password):
+    try:
+        server_uri = config.LDAP_URL
+        server = Server(server_uri, get_info=ALL)
+        connection = Connection(server,
+                        user='cn='+user_id+',ou=users,'+config.LDAP_ADMIN_DN,
+                        password=password)
+        bind_response = connection.bind()
+    except LDAPBindError as e:
+        bind_response = e
+    return bind_response
